Add files via upload
File for printing Masks. TPU or other Flex filament is recommended for the mask and PLA or another rigid filament is recommended for the filter adapter. See the Readme file for details. There are three different size masks with engraved size letters on the faceplate of the masks. The filter adapter will need to be scaled according to the mask size being printed. TPU filament can be found here: https://www.matterhackers.com/store/c/TPU%20%28Thermoplastic%20Polyurethane%29